requirements.txt
Flask==2.0.2
python-dotenv==0.19.1
slackclient==2.9.3
slackeventsapi==3.0.0
